+++ CHANGES.md
@@ -91,6 +91,13 @@ the versioning.
      -  Added `@fedify/nestjs` package.
      -  Added `FedifyModule` for integrating Fedify into NestJS applications.
 
+ -  Added `SqliteKvStore`, implementing `KvStore` using SQLite with the
+    `@fedify/sqlite` package. Compatible with Bun, Deno, and Node.js.
+    [[#274], [#318] by An Subin]
+
+     -  Added `@fedify/sqlite` package.
+     -  Added `SqliteKvStore`, a SQLite implementation of `KvStore`.
+
  -  Added `-o`/`--output` option to `fedify lookup` command. This option allows
     users to save retrieved lookup results to specified path.
     [[#261], [#321] by Jiwon Kwon]
@@ -140,6 +147,8 @@ the versioning.
 [#298]: https://github.com/fedify-dev/fedify/pull/298
 [#304]: https://github.com/fedify-dev/fedify/issues/304
 [#309]: https://github.com/fedify-dev/fedify/pull/309
+[#274]: https://github.com/fedify-dev/fedify/issues/274
+[#318]: https://github.com/fedify-dev/fedify/pull/318
 [#310]: https://github.com/fedify-dev/fedify/issues/310
 [#311]: https://github.com/fedify-dev/fedify/issues/311
 [#321]: https://github.com/fedify-dev/fedify/pull/321
@@ -166,6 +175,7 @@ Released on July 28, 2025.
 
 [#300]: https://github.com/fedify-dev/fedify/pull/300
 [#335]: https://github.com/fedify-dev/fedify/pull/335
+
 
 
 Version 1.7.6